requirements.txt
dicttoxml==1.7.4
EasyConnect==1.0.5
LiteSpeed~=1.0.8
pathtools==0.1.2
Pillow~=7.2.0
pystache==0.5.4
watchdog==0.10.3

pdf2image~=1.13.1
future~=0.18.2